Added option for keeping tmp files from scenario.

+++ cleanup.sh
@@ -3,4 +3,7 @@
 sudo pkill -9 vcmd
 sudo pkill -9 python
 sudo core-cleanup
-sudo rm -rf /tmp/*_server.csv /tmp/*_client.csv /tmp/.auto-scenario.log /tmp/serval* /tmp/netmon-n* /tmp/n* /tmp/pf_* /tmp/mf_*
+
+if [[ "$1" == "a" ]]; then
+    sudo rm -rf /tmp/*_server.csv /tmp/*_client.csv /tmp/.auto-scenario.log /tmp/serval* /tmp/netmon-n* /tmp/n* /tmp/pf_* /tmp/mf_*
+fi
